requirements.txt
# Core
streamlit>=1.40.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# ML / Forecasting
prophet>=1.1.5
statsmodels>=0.14.0
scikit-learn>=1.3.0

# AI Summaries
google-genai>=0.1.0

# Firebase
firebase-admin>=6.2.0

# Data handling
openpyxl>=3.1.0
python-dotenv>=1.0.0
